Fix failing tests due to restructuring of folders
